+++ R/aaa.R
@@ -51,7 +51,8 @@ globalVariables(
     "pre_denoise_isotype_fraction", "umis_denoised", "isotype_reduction",
     "fraction_denoised", "median_isotype_reduction",
     "median_pre_denoise_isotype", "median_post_denoise_isotype",
-    "x_pos", "y_pos"
+    "x_pos", "y_pos", "umi_type", "n_outliers", "max_degree",
+    "outlier_label"
   ),
   package = "pixelatorES",
   add = TRUE